Conflicts:
	README.md
	modules/clientBinaries/generateVersions.js
	modules/init.js
	modules/notification/notification.js
	package.json
	src/app/core-ui/core-ui.module.ts
	src/app/core-ui/main/cart/cart.component.html
	src/app/core-ui/main/cart/cart.component.ts
	src/app/core-ui/main/main-view.component.html
	src/app/core-ui/main/main-view.module.ts
	src/app/core/market/api/bid/bid.model.ts
	src/app/core/market/api/bid/bid.service.ts
	src/app/core/market/api/cart/cart.model.ts
	src/app/core/market/api/category/category.service.ts
	src/app/core/market/api/listing/listing.model.ts
	src/app/core/market/api/listing/listing.service.ts
	src/app/core/market/api/profile/address/address.service.ts
	src/app/core/market/api/template/image/image.service.ts
	src/app/core/market/api/template/image/imagecollection.model.ts
	src/app/core/market/api/template/template.model.ts
	src/app/core/market/api/template/template.service.ts
	src/app/core/market/market-cache/add-to-cart-cache.service.ts
	src/app/core/market/market-cache/post-listing-cache.service.ts
	src/app/core/market/market-state/market-state.service.ts
	src/app/core/market/market.module.ts
	src/app/core/notification/notification.service.ts
	src/app/core/util/utils.ts
	src/app/market/_lists-shared.scss
	src/app/market/buy/buy.component.html
	src/app/market/buy/buy.component.scss
	src/app/market/buy/checkout-process/checkout-process.component.html
	src/app/market/buy/checkout-process/checkout-process.component.ts
	src/app/market/listings/listing-item/listing-item.component.html
	src/app/market/listings/listing-item/listing-item.component.ts
	src/app/market/listings/listings.component.html
	src/app/market/listings/listings.component.scss
	src/app/market/listings/listings.component.spec.ts
	src/app/market/listings/listings.component.ts
	src/app/market/listings/preview-listing/preview-listing.component.html
	src/app/market/listings/preview-listing/preview-listing.component.scss
	src/app/market/listings/preview-listing/preview-listing.component.spec.ts
	src/app/market/listings/preview-listing/preview-listing.component.ts
	src/app/market/market.module.ts
	src/app/market/sell/add-item/add-item.component.html
	src/app/market/sell/add-item/add-item.component.scss
	src/app/market/sell/add-item/add-item.component.ts
	src/app/market/sell/sell.component.html
	src/app/market/sell/sell.component.scss
	src/app/market/sell/sell.component.spec.ts
	src/app/market/sell/sell.component.ts
	src/app/market/sell/seller-listing/seller-listing.component.html
	src/app/market/sell/seller-listing/seller-listing.component.ts
	src/app/market/sell/status.class.ts
	src/app/market/shared/favorite/favorite.component.html
	src/app/market/shared/favorite/favorite.component.scss
	src/app/market/shared/orders/order-item/order-item.component.html
	src/app/market/shared/orders/order-item/order-item.component.ts
	src/app/market/shared/orders/orders.component.html
	src/app/market/shared/orders/orders.component.scss
	src/app/market/shared/orders/orders.component.ts
	src/app/modals/_confirmations-shared.scss
	src/app/modals/market-bid-confirmation-modal/bid-confirmation-modal.component.html
	src/app/modals/market-bid-confirmation-modal/bid-confirmation-modal.component.spec.ts
	src/app/modals/market-bid-confirmation-modal/bid-confirmation-modal.component.ts
	src/app/modals/market-place-order/place-order.component.html
	src/app/modals/market-place-order/place-order.component.scss
	src/app/modals/market-place-order/place-order.component.spec.ts
	src/app/modals/market-place-order/place-order.component.ts
	src/app/modals/market-shipping/shipping.component.html
	src/app/modals/market-shipping/shipping.component.scss
	src/app/modals/market-shipping/shipping.component.spec.ts
	src/app/modals/market-shipping/shipping.component.ts
	src/app/modals/modals.module.ts
	src/app/wallet/wallet/send/send.component.html
	yarn.lock

+++ modules/daemon/daemon.js
@@ -5,7 +5,7 @@ const rxIpc = require('rx-ipc-electron/lib/main').default;
 const Observable = require('rxjs/Observable').Observable;
 
 const _options = require('../options');
-const clearCookie = require('../webrequest/http-auth').removeWalletAuthentication;
+const removeWalletAuthentication = require('../webrequest/http-auth').removeWalletAuthentication;
 const rpc = require('../rpc/rpc');
 const cookie = require('../rpc/cookie');
 const daemonManager = require('../daemon/daemonManager');
@@ -25,8 +25,9 @@ exports.init = function () {
     return Observable.create(observer => {
       console.log('got data on daemon channel!');
       if (data && data.type === 'restart') {
-        exports.restart(true);
-        observer.complete(true);
+        exports.restart(true).then(() => {
+          observer.complete(true);
+        });
       } else {
         observer.complete(true);
       }
@@ -36,27 +37,35 @@ exports.init = function () {
 
 exports.restart = function (alreadyStopping) {
   log.info('restarting daemon...')
+  return (new Promise((resolve, reject) => {
+    // setup a listener, waiting for the daemon
+    // to exit.
+    if (daemon) {
+      daemon.once('close', code => {
+        log.info('clearing cookie, encrypt wallet')
+        // clear authentication
+        removeWalletAuthentication();
 
-  // setup a listener, waiting for the daemon
-  // to exit.
-  if (daemon) {
-    daemon.once('close', code => {
-      // clear authentication
-      clearCookie();
-      
-      // restart
-      this.start(chosenWallets);
-    });
-  }
+        // restart
+        this.start(chosenWallets);
+        resolve();
+      });
+    } else {
+      // daemon not in our control
+      resolve();
+    }
 
-  // wallet encrypt will restart by itself
-  if (!alreadyStopping) {
-    // stop daemon but don't make it quit the app.
-    const restarting = true;
-    exports.stop(restarting).then(() => {
-      log.debug('waiting for daemon shutdown...')
-    });
-  }
+    // wallet encrypt will restart by itself
+    if (!alreadyStopping) {
+      // stop daemon but don't make it quit the app.
+      const restarting = true;
+      exports.stop(restarting).then(() => {
+        log.debug('waiting for daemon shutdown...')
+      });
+    }
+  }));
+
+
 
 }
 
@@ -66,10 +75,16 @@ exports.start = function (wallets) {
     chosenWallets = wallets;
 
     exports.check().then(() => {
-      log.info('daemon already started');
+      log.info('daemon already running');
       resolve(undefined);
 
-    }).catch(() => {
+    }).catch(async () => {
+      // clear cookie authentication
+      // currently used by electron (if reboot)
+      removeWalletAuthentication();
+      await askForDeletingCookie().catch(() => log.info('cookie: already existed and re-using.'));
+
+      daemon = undefined;
 
       let options = _options.get();
       const daemonPath = options.customdaemon
@@ -81,21 +96,15 @@ exports.start = function (wallets) {
 
       const child = spawn(daemonPath, [...process.argv, "-rpccorsdomain=http://localhost:4200", ...wallets])
         .on('close', code => {
-          log.info('daemon exited - setting to undefined.');
-          daemon = undefined;
           if (code !== 0) {
             reject();
             log.error(`daemon exited with code ${code}.\n${daemonPath}\n${process.argv}`);
-          } else {
-            log.info('daemon exited successfully');
           }
         });
 
       // TODO change for logging
       child.stdout.on('data', data => daemonData(data, console.log));
-      child.stderr.on('data', data => {
-        daemonData(data, console.log);
-      });
+      child.stderr.on('data', data => daemonData(data, console.log));
 
       daemon = child;
     });
@@ -120,6 +129,8 @@ exports.check = function () {
   });
 }
 
+// Note: this will resolve before the daemon has actually quit.
+// do not rely on it. Use daemon.on('close', ...) instead
 exports.stop = function (restarting) {
   log.info('daemon stop called..');
   return new Promise((resolve, reject) => {
@@ -130,26 +141,64 @@ exports.stop = function (restarting) {
       // do not close electron when restarting (e.g encrypting wallet)
       if (!restarting) {
         daemon.once('close', code => {
-          log.info('daemon exited successfully - we can now quit electron safely! :)');
+          log.info('daemon exited successfully - quiting electron');
           electron.app.quit();
         });
       }
 
       log.info('Call RPC stop!');
-      rpc.call('stop', null, (error, response) => {
-        if (error) {
-          log.info('daemon errored to rpc stop - killing it brutally :(');
-          daemon.kill('SIGINT');
-          reject();
-        } else {
-          log.info('Daemon stopping gracefully...');
-          resolve();
-        }
-      });
+      _stop();
+      resolve();
+
     } else {
       log.info('Daemon not managed by gui.');
+      electron.app.quit();
       resolve();
     }
 
+  });
+}
+
+function _stop(attempt = 0) {
+  rpc.call('stop', null, (error, response) => {
+    if (error) {
+      log.info('daemon errored - trying again in a second..')
+      // just kill after 180s
+      if (attempt >= 180) {
+        log.info('daemon errored to rpc stop - killing it brutally :(');
+        log.info(error);
+        daemon.kill('SIGINT');
+      } else {
+        // daemon is busy cut it some slack.
+        setTimeout(_stop, 1000, ++attempt);
+      }
+    } else {
+      log.info('Daemon stopping gracefully - we can now quit safely :)');
+    }
+  });
+}
+
+function askForDeletingCookie() {
+  return new Promise((resolve, reject) => {
+    if (cookie.checkCookieExists()) {
+      // alert for cookie
+      log.info('cookie: dialog open');
+      electron.dialog.showMessageBox({
+        type: 'warning',
+        buttons: ['Yes', 'No'],
+        message: `It seems like you already have an instance of Particl running, do you want to connect to that instead? 
+                  If you think you're having issues starting the application, select no.`
+      }, (response) => {
+        if (response === 1) {
+          log.info('cookie: deleting');
+          cookie.clearCookieFile();
+          resolve();
+        } else {
+          reject();
+        }
+      });
+    } else {
+      resolve();
+    }
   });
 }
